- Added Simulatoin Class - Added side counters
