+++ Counters/HL_Counter.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BlackJack.CountingStrategy
+namespace BlackJackSimul.CountingStrategy
 {
     public class HL_Counter : ACounter
     {
@@ -16,12 +16,22 @@ namespace BlackJack.CountingStrategy
             switch (cardValue)
             {
                 case "A":
+                    {
+                        increment = -1;
+                        AceCounter++;
+                        break;
+                    }
+                    
                 case "10":
                 case "J":
                 case "Q":
                 case "K":
-                    increment = -1;
-                    break;
+                    {
+                        increment = -1;
+                        TenCounter++;
+                        break;
+                    }
+                    
                 case "2":
                 case "3":
                 case "4":
@@ -37,5 +47,13 @@ namespace BlackJack.CountingStrategy
             return increment;
             
        }
+
+        public override void UpdateCounters(string cardFace)
+        {
+            var RemainingDecks = (float)_shoe.cards.Count / (float)Costanti.N_CARTE_MAZZO;
+            RunningCounter += CalulateIncrement(cardFace);
+            TrueCounter = ((float)RunningCounter / RemainingDecks);
+            UpddateSideCounters(cardFace);
+        }
     }
 }
