+++ Data/PlayRecord.cs
@@ -1,21 +1,19 @@
-﻿using BlackJackSimul.CountingStrategy;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BlackJackSimul
+namespace BlackJackSimul.Data
 {
-    public class SingleExstractionData
+    public class PlayRecord
     {
-        public int ShoeNumber { get; set; } = 1;
-        public int deckNumber { get; set; } = 1;
-        public string Card { get; set; }
+        public int ShoeID{ get; set; } = 0;
+        public int PlayID { get; set; } = 0;
+        public string CardSequence { get; set; }
         public int HL_RunningCounter { get; set; }
         public float HL_TrueCounter { get; set; }
         public int RAPC_RunningCounter { get; set; }
         public float RAPC_TrueCounter { get; set; }
-      
     }
 }
