+++ Program.cs
@@ -1,19 +1,15 @@
-﻿using BlackJack.CountingStrategy;
+﻿using BlackJackSimul.CountingStrategy;
+using BlackJackSimul.Data;
 using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text;
 
-namespace BlackJack
+namespace BlackJackSimul
 {
     class Program
     {
-        static int numeroShoe = 0;
-        static int numeroGiocate = 0;
-        static Log log = new Log();
-       
-
         static void Main(string[] args)
         {
             // Single extraction data
@@ -72,244 +68,10 @@ namespace BlackJack
 
             WriteStatistic(stat_data);
 
-
-            //FileCSV handFile = new FileCSV("HandFile");
-            //var had_data = new HandData
-            for (int nShoes = 1; nShoes <= Costanti.N_SHOES; nShoes++)
-            {
-                Shoe testShoe = new Shoe(Costanti.N_MAZZI_PER_SHOE, Costanti.N_MAZZI_DA_ESTRARRE_PER_SHOE);
-                testShoe.Shuffle();
-                SimulatePlay(testShoe.cards);
-            }
-                     
-        }
-
-
-        private static void SimulatePlay(Queue<string> cardSequence)
-        {
-            numeroShoe++;
-
-            while (cardSequence.Count >= 208   )
-            {
-                numeroGiocate++;
-
-                #region DISTRIBUZIONE INIZIALE
-
-                var player = new Player();
-                player.NewHand();
-                var dealer = new Dealer();
-                dealer.NewHand();
-
-                //Estrazione delle prime quattro carte
-                var playerFirstCard = cardSequence.Dequeue();
-                var dealerFirstCard = cardSequence.Dequeue();
-                var playerSeconCard = cardSequence.Dequeue();
-                var dealerSeconCard = cardSequence.Dequeue();
-
-                //Carta al player
-                player.GiveCard(playerFirstCard);
-
-                //Carta al dealer
-                dealer.GiveCard(dealerFirstCard);
-
-                //Carta al player
-                player.GiveCard(playerSeconCard);
-
-                //Carta al dealer
-                dealer.GiveCard(dealerSeconCard);
-                #endregion DISTRIBUZIONE INZIZIALE
-           
-                #region GIOCATA PLAYER
-                
-              //  if(Costanti.f_console)
-           //     {
-                    Console.ForegroundColor = ConsoleColor.White;
-                    log.Write($"\nShoe {numeroShoe} \t Mano {numeroGiocate} \t");
-                    log.WriteLine($"DEALER: { dealerFirstCard}");
-           //     }
-
-
-                //Giocata player
-                if (!Util.CheckBlackJack(dealer.hand))
-                {
-                    int dealerFirstCardPoint = Util.PointOf(dealerFirstCard);
-                    for (int playerhandID = 0; playerhandID < player.hands.Count; playerhandID++)
-                    {
-                        var playerHand = player.hands[playerhandID];
-                        string response = "";
-
-                        if(Costanti.f_console)
-                            if (playerhandID > 0)
-                                log.WriteLine("---");
-                        player.WriteHandResult(playerHand);
-
-                        while (!playerHand.f_bust &&
-                                response != "STAND" &&
-                                !playerHand.f_double)
-                        {
-                            response = player.Ask(playerhandID, dealerFirstCardPoint);
-
-                            if (response == "HIT")
-                                player.GiveCard(cardSequence.Dequeue(), playerhandID);
-
-                            else if (response == "DOUBLE DOWN")
-                            {
-                                //Check del double
-                                if (playerHand.Cards.Count == 2)
-                                {
-                                    playerHand.Puntata *= 2;
-                                    playerHand.f_double = true;
-                                }
-                                //Se non si può raddoppiare
-                                player.GiveCard(cardSequence.Dequeue(), playerhandID);
-                            }
-                            else if (response == "SPLIT")
-                            {
-                                //Controllo aggiuntivo
-                                if (playerHand.Cards.Count == 2 && playerHand.f_coppia)
-                                {
-                                    //Creo nuova mano e sposto la carta da una mano all'altra
-                                    player.NewHand(true);
-                                    //Setto il flag split
-                                    playerHand.f_split = true;
-                                    player.hands[playerhandID + 1].f_split = true;
-                                   
-                                    player.GiveCard(playerHand.Cards[1], playerhandID + 1);
-                                    playerHand.Cards.RemoveAt(playerHand.Cards.Count - 1);
-                                    var firstHandSecondCard = cardSequence.Dequeue();
-                                    var secondHandSecondCard = cardSequence.Dequeue();
-                                    player.GiveCard(firstHandSecondCard, playerhandID);
-                                    player.GiveCard(secondHandSecondCard, playerhandID+1);
-                                    
-                                }
-                            }
-
-                            if(Costanti.f_console)
-                            {
-                                log.WriteLine(response);
-                                if (response == "SPLIT")
-                                    log.WriteLine("---");
-                            }
-                           
-                            player.WriteHandResult(playerHand);
-
-                        }
-                     }
-
-                    #endregion GIOCATA PLAYER
-
-
-                    #region GIOCATA DEALER
-
-                    //Controllo che il player non abbia fatto tutti blackjack o bust
-                    int countEnded = 0;
-                    foreach (Hand playerHand in player.hands)
-                    {
-                        if (Util.CheckBlackJack(playerHand))
-                            countEnded++;
-                    }
-
-
-                    string dealerAction = "";
-                    if (countEnded != player.hands.Count)
-                        while (dealer.hand.punteggio.Value < 21 &&
-                               !dealer.hand.f_bust &&
-                               dealerAction != "STAND")
-                        {
-                            dealer.WriteResult();
-                            dealerAction = dealer.ApplicaRegole();
-                            log.WriteLine(dealerAction);
-                            if (dealerAction == "HIT")
-                                dealer.GiveCard(cardSequence.Dequeue());
-
-                        }
-
-                    dealer.WriteResult();
-
-                    #endregion GIOCATA DEALER
-                }
-                else
-                {
-                    player.WriteResult();
-                    dealer.WriteResult();
-                }
-
-                //Risultato della mano
-                CheckTheWinner(player, dealer);
-
-              //  Console.ReadKey();     
-              
-            }
-        }
-
-        static void CheckTheWinner(Player player, Dealer dealer)
-        {
-            Console.ForegroundColor = ConsoleColor.Green;
-            var dealerHand = dealer.hand;
-            var punteggioDealer = dealerHand.punteggio.Value;
-
-                foreach (Hand playerHand in player.hands)
-                {
-                    var punteggioPlayer = playerHand.punteggio.Value;
-
-                    //Risultato partita
-                    if (punteggioPlayer > 21)
-                    {
-                        Player.ManiVinte--;
-                        Player.TotVincita -= playerHand.Puntata;
-                        log.WriteLine($"Il dealer vince");
-                }
-
-                    else if (punteggioDealer > 21)
-                    {
-                        Player.ManiVinte++;
-                        Player.TotVincita += playerHand.Puntata;
-                        log.WriteLine($"Il player vince");
-                    }
-
-                    else if (Util.CheckBlackJack(playerHand))
-                    {
-                        Player.CounterBlackJack++;
-                        if (!Util.CheckBlackJack(dealerHand))
-                        {
-                            Player.ManiVinte++;
-                            Player.TotVincita += (float)playerHand.Puntata * (float)1.5;
-                            log.WriteLine($"Il player vince");
-                        }
-                        else
-                        {
-                            Dealer.CounterBlackJack++;
-                        if (Costanti.f_console)
-                            log.WriteLine($"Push");
-                        }
-                    }
-
-                    else if (punteggioPlayer > punteggioDealer)
-                    {
-                        Player.ManiVinte++;
-                        Player.TotVincita += playerHand.Puntata;
-                        log.WriteLine($"Il player vince");
-                    }
-
-                    else if (punteggioPlayer < punteggioDealer)
-                    {
-                        Player.ManiVinte--;
-                        Player.TotVincita -= playerHand.Puntata;
-                        log.WriteLine($"Il dealer vince");
-                    }
-
-                    else
-                    {
-                        Player.ManiPareggiate++;
-                        log.WriteLine($"Push");
-                    }
-                }
-            
-            Console.ResetColor();
-            log.WriteLine($"Saldo player: {Player.TotVincita}");
+            Simulation simulation = new Simulation(Costanti.N_SHOES, Costanti.N_MAZZI_PER_SHOE, Costanti.N_MAZZI_DA_ESTRARRE_PER_SHOE);
+            simulation.Start();
 
         }
-
 
 
         private static void WriteStatistic(StatisticData stat_data)
